introduced loader factory to make it simpler to create in process and out of process loader processing
