+++ src/loader.js
@@ -1,12 +1,17 @@
-var path = require("path");
+var Bitloader = require("bit-loader");
 var utils = require("belty");
 var resolvePath = require("bit-bundler-utils/resolvePath");
 var readFile = require("bit-bundler-utils/readFile");
 var pluginLoader = require("./pluginLoader");
 var logger = require("./logger").create("bundler/loader");
 
+var moduleNotFoundError = buildError.bind(null, "Unable to find module");
+var moduleNotLoadedError = buildError.bind(null, "Unable to load module");
+var moduleNotResolvedError = buildError.bind(null, "Unable to resolve module");
 
 function Loader(options) {
+  options = options || {};
+
   Bitloader.call(this, utils.extend({}, options, {
     resolve: configureResolve(options),
     fetch: configureFetch(options),
@@ -14,98 +19,58 @@ function Loader(options) {
   }));
 }
 
-Loader.prototype.init = function() {
-  return init(this, {
-    type: "init",
-    data: this.options
-  });
-};
-
-Loader.prototype.fetch = function(name, referrer) {
-  var loader = this;
-  return send(this, {
-    type: "fetch",
-    data: {
-      name: name,
-      referrer: referrer
-    }
-  })
-  .then(function(result) {
-    loader.cache = result.cache;
-    return result.module;
-  });
-};
-
-Loader.prototype.getModule = function(id) {
-  return this.cache[id];
-};
-
-Loader.prototype.deleteModule = function(mod) {
-  delete this.cache[mod.id];
-  send(this, {
-    type: "delete",
-    id: mod.id
-  });
-};
+Loader.prototype = Object.create(Bitloader.prototype);
+Loader.prototype.constructor = Loader;
 
 Loader.prototype.getCache = function() {
   return this.cache;
 };
 
+function configureResolve(options) {
+  var resolver = resolvePath.configure({baseUrl: options.baseUrl});
 
-function init(loader, message) {
-  if (loader._init) {
-    return loader._init.deferred;
-  }
+  return function resolveName(meta) {
+    function handleError(err) {
+      var message = err && err.message;
 
-  var id = loader.id++;
-  var deferred = new Promise(function(resolve, reject) {
-    loader._init = { id: id };
-    loader._init.resolve = resolve;
-    loader._init.reject = reject;
-    loader._init.deferred = deferred;
-    loader.loaderProc.send(utils.assign({ id: id }, message));
-  });
-
-  return deferred;
-}
-
-function send(loader, message) {
-  var id = loader.id++;
-  var deferred = new Promise(function(resolve, reject) {
-    loader.pending[id] = {};
-    loader.pending[id].resolve = resolve;
-    loader.pending[id].reject = reject;
-    loader.pending[id].deferred = deferred;
-
-    loader.init().then(function() {
-      return loader.loaderProc.send(utils.assign({ id: id }, message));
-    });
-  });
-
-  return deferred;
-}
-
-function registerLoaderProcHandlers(loader) {
-  loader.loaderProc
-    .on("error", function(error) {
-      console.error(error);
-    })
-    .on("message", function(message) {
-      if (loader.pending.hasOwnProperty(message.id)) {
-        handleResult(message, loader.pending[message.id]);
-        delete loader.pending[message.id];
+      if (message && message.indexOf("Cannot find module") === 0 && options.ignoreNotFound) {
+        logger.warn("Module not found. Skipping it.", moduleNotFoundError(meta));
       }
-      else if (loader._init.id === message.id) {
-        handleResult(message, loader._init);
+      else {
+        logger.error(moduleNotResolvedError(meta), err);
+        throw err;
       }
-    });
+
+      return {
+        id: "@notfound",
+        path: ""
+      };
+    }
+
+    return resolver(meta).then(utils.identity, handleError);
+  };
 }
 
-function handleResult(message, pending) {
-  message.error ?
-    pending.reject(message.error) :
-    pending.resolve(message.data);
+function configureFetch(options) {
+  return function fetchModule(meta) {
+    function handleError(err) {
+      if (err && err.code === "ENOENT" && options.ignoreNotFound) {
+        logger.warn("Module not found. Skipping it.", moduleNotFoundError(meta));
+        return Promise.resolve({ source: "module.exports = require('" + meta.name + "')" });
+      }
+
+      logger.error(moduleNotLoadedError(meta), err);
+      throw err;
+    }
+
+    return meta.id === "@notfound" && options.ignoreNotFound ?
+      Promise.resolve({ source: "" }) :
+      readFile(meta).then(utils.identity, handleError);
+  };
+}
+
+function buildError(title, meta) {
+  return title + " '" + meta.name + "'." + (meta.referrer ? " Referrer " + JSON.stringify(meta.referrer.path) : "");
 }
 
 module.exports = Loader;
